code refactor to improve flexibility of plotting
[NEW] added a curve smoothing module to simplify the plotting of smoothed
    counts data
[CHANGED] the GeneIndex class and factory function moved to new module,
    gene_data
[CHANGED] old get_gene_data functionality now implemented in
    rdump_to_stable_ids

The idea is we use rdump_to_stable_ids on the gene data from Rohan and generate
a list of Ensembl Stable IDs. These are then used by the segment_count script,
which imports the GetGeneIndexes factory function. This will allow us to check
the results for arbitrary lists of genes, defined outside the R-process.